[STD-19] Completing transformation structure and world tiles

+++ src/common/inspection/Inspectable.cpp
@@ -45,7 +45,7 @@ namespace Metal {
         return fields;
     }
 
-    int Inspectable::getChangeId() const {
+    unsigned long Inspectable::getChangeId() const {
         return changes;
     }
